uint32 replaced uint8

+++ Transaction.sol
@@ -3,18 +3,18 @@ pragma solidity ^0.4.11;
 
 contract Transaction {
   
-  mapping (bytes32 => uint8) public candidateCredit;
+  mapping (bytes32 => uint32) public candidateCredit;
   
   bytes32[] public candidateList;
   function Transaction(bytes32[] candidateNames) {
     candidateList = candidateNames;
   }
 
-  function getCredit(bytes32 candidate) returns (uint8) {
+  function getCredit(bytes32 candidate) returns (uint32) {
     return candidateCredit[candidate];
   }
 
-  function sendMoney(bytes32 candidate, uint8 money) {
+  function sendMoney(bytes32 candidate, uint32 money) {
       candidateCredit[candidate] += money;
   }
